27pseudo

+++ exercise26/src/main/java/PaymentCalculator.java
@@ -3,7 +3,7 @@ import java.lang.Math;
 public class PaymentCalculator {
     private final double BALANCE, MONTHLY_PAYMENT, DAILY_RATE;
 
-    PaymentCalculator(int balance, int apr, int monthlyPayment){
+    PaymentCalculator(double balance, double apr, double monthlyPayment){
         BALANCE = balance;
         DAILY_RATE = apr / 100.0 / 365.0;
         MONTHLY_PAYMENT = monthlyPayment;
